+++ fedmsg_meta_fedora_infrastructure/centos_ci.py
@@ -41,99 +41,115 @@ class CentosCiProcessor(BaseProcessor):
         namespace = msg['msg']['namespace']
         status = msg['msg']['status']
 
-        if 'ci.pipeline.package.ignore' in msg['topic']:
-            tmpl = self._(
-                'Commit "{commit}" of package {ns}/{pkg} is being '
-                'ignored by the CI pipeline on branch {branch}')
-
-        elif 'ci.pipeline.complete' in msg['topic']:
+        def _get_status(status):
             if status.lower() == 'success':
                 status = 'passed'
             elif status.lower() == 'aborted':
                 status = 'was aborted on'
             else:
                 status = 'failed'
+            return status
+
+        if 'ci.pipeline.package.ignore' in msg['topic']:
+            tmpl = self._(
+                'Commit "{commit}" of package {ns}/{pkg} is being '
+                'ignored by the Atomic CI pipeline on branch {branch}')
+
+        elif 'ci.pipeline.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
                 'Commit "{commit}" of package {ns}/{pkg} {status}'
-                ' the CI pipeline on branch {branch}')
+                ' the Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.package.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed its build '
-                'in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} building '
+                'in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.package.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued to be built '
-                'in the CI pipeline on branch {branch}')
+                'in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.package.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is being built in the '
-                'CI pipeline on branch {branch}')
+                'Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.package.test.functional.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed its '
-                'functional tests in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} its '
+                'functional tests in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.package.test.functional.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued for '
-                'functional testing in the CI pipeline on branch {branch}')
+                'functional testing in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.package.test.functional.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is running its '
-                'functional tests in the CI pipeline on branch {branch}')
+                'functional tests in the Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.compose.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed a '
-                'compose in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} a '
+                'compose in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.compose.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued for a '
-                'compose in the CI pipeline on branch {branch}')
+                'compose in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.compose.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is being part of a '
-                'compose in the CI pipeline on branch {branch}')
+                'compose in the Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.compose.test.integration.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed tests as '
-                'part of a compose in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} its tests as '
+                'part of a compose in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.compose.test.integration.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued for tests '
-                'as part of a compose in the CI pipeline on branch {branch}')
+                'as part of a compose in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.compose.test.integration.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is being tested as '
-                'part of a compose in the CI pipeline on branch {branch}')
+                'part of a compose in the Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.image.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed being built '
-                'in an image in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} being built '
+                'in an image in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.image.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued to be built '
-                'in an image in the CI pipeline on branch {branch}')
+                'in an image in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.image.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is being built '
-                'in an image in the CI pipeline on branch {branch}')
+                'in an image in the Atomic CI pipeline on branch {branch}')
 
         elif 'ci.pipeline.image.test.smoke.complete' in msg['topic']:
+            status = _get_status(status)
             tmpl = self._(
-                'Commit {commit} of package {ns}/{pkg} completed tests '
-                'in an image in the CI pipeline on branch {branch}')
+                'Commit {commit} of package {ns}/{pkg} {status} its tests '
+                'in an image in the Atomic CI pipeline on branch {branch}')
+
         elif 'ci.pipeline.image.test.smoke.queued' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is queued to be tested '
-                'in an image in the CI pipeline on branch {branch}')
+                'in an image in the Atomic CI pipeline on branch {branch}')
         elif 'ci.pipeline.image.test.smoke.running' in msg['topic']:
             tmpl = self._(
                 'Commit {commit} of package {ns}/{pkg} is being tested '
-                'in an image in the CI pipeline on branch {branch}')
+                'in an image in the Atomic CI pipeline on branch {branch}')
 
         else:
             tmpl = ""